refactor(gui): extract config tests to sibling file
primer-gui/src/config.rs was 1985 lines, dominated by a ~1085-line
`#[cfg(test)] mod tests` (lines 900-1985). Move that module verbatim
into a new sibling config/tests.rs using the Rust-2018
config.rs + config/tests.rs subdirectory form — same zero-churn pattern
as PRs #268/#269/#270/#271.

config.rs drops 1985 -> 901 lines (-1084); it now holds only the
production serde DTOs, defaults, and load/save helpers plus a 2-line
`#[cfg(test)] mod tests;` declaration. The test module reaches every
private parent item via `use super::*`, so no visibility changed.

Pure test-only relocation: no behaviour change, no incidental fixes.

Verification:
- config tests: 56 passed (default) / 56 (--no-default-features) — baseline 56
- workspace: 0 failures (51 "test result: ok")
- clippy: clean (default workspace + primer-gui --no-default-features)
- fmt: clean

The moved span contains one `#[cfg(feature = "embedding")]` test, so the
embedding-on (default) and embedding-off (--no-default-features) clippy +
test matrix was run to confirm the cfg gate compiles in both states.
